+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Refactor all classes and interfaces to client side only
     - Remove meteor/check dependency
     - Introduce new Forms.C.CheckStatus.TRANSPARENT check status display constant
+    - package.js refactoring
 
 ### 1.2.0